(Refactoring) CurrencyUtil moved to the PresentationLayer for using resource files.
