+++ presentationlayer/src/main/java/ru/project/worldmoneyinfo/ui/statistic_screen/StatisticViewModel.java
@@ -17,7 +17,7 @@ import io.reactivex.schedulers.Schedulers;
 import ru.project.domainlayer.model.StatisticCurrencyData;
 import ru.project.domainlayer.service.ICurrenciesService;
 import ru.project.domainlayer.service.ISettingsService;
-import ru.project.domainlayer.utils.CurrencyUtil;
+import ru.project.worldmoneyinfo.utils.CurrencyUtil;
 import ru.project.domainlayer.utils.DateUtil;
 import ru.project.worldmoneyinfo.MainApplication;
